Update for the MediaTracker section
Added a "What is the MediaTracker ?" section
Added somes contents of the Wiki.
I wanted to add a 3rd step for shooting videos or creating screenshots but didn't had the time to do that.
Sorry if i made somes mistakes. I tried to do a clear tutorial. I hope it's the case.